Add Job class for composition
