+++ lab3/prob1/Job.java
@@ -1,0 +1,11 @@
+package lesson3.labs.prob1;
+
+public class Job {
+    private double salary;
+    public Job(double salary) {
+        this.salary = salary;
+    }
+    public double getSalary() {
+        return salary;
+    }
+}
