+++ lab3/prob1/Person.java
@@ -1,25 +1,32 @@
 package lesson3.labs.prob1;
 
-
 public class Person {
-
 	private String name;
-	Person(String n) {
+	private Job job;
+	public Person(String n) {
 		name = n;
+	}
+	public Person(String n, double salary) {
+		name = n;
+		this.job = new Job(salary);
 	}
 	public String getName() {
 		return name;
 	}
 	@Override
 	public boolean equals(Object aPerson) {
-		if(aPerson == null) return false; 
+		if(aPerson == null) return false;
 		if(!(aPerson instanceof Person)) return false;
 		Person p = (Person)aPerson;
 		boolean isEqual = this.name.equals(p.name);
 		return isEqual;
 	}
 	public static void main(String[] args) {
-		
+		Person p1 = new Person("Joe", 30000);
+		Person p2 = new Person("Joe");
+		//As PersonsWithJobs, p1 should be equal to p2
+		System.out.println("p1.equals(p2)? " + p1.equals(p2));
+		System.out.println("p2.equals(p1)? " + p2.equals(p1));
 	}
-
 }
+
